refactor(resources): extract shared helpers into resources/common
Phase 5 of the architecture review deduplication plan:

- Add resources/common with resource_id(), belongs_to(), song_thing(),
  blob_thing(), player_from_song_links(), SongLinkRecord, FetchedSongRecord
- Remove 4 identical *_resource() functions (setlist, collection, song, blob)
- Remove 4 identical *_belongs_to() free functions and inline belongs_to()
- Remove duplicate song_thing/blob_thing from collection and song modules
- Unify player_from_song_links / collection_player_from_links (same logic)
- Consolidate SongLinkRecord and FetchedSongRecord (duplicated setlist+collection)
- Move resource_id and belongs_to tests into common::tests

Made-with: Cursor
