+++ backend/src/resources/mod.rs
@@ -1,5 +1,7 @@
 pub mod rest;
 
+pub(crate) mod common;
+
 pub mod blob;
 pub use blob::{Blob, CreateBlob};
